changed to Spring Security, created JWT token for authentication, created UsersDTO

+++ src/main/java/com/example/ecommercebackend/service/CartService.java
@@ -1,11 +1,13 @@
 package com.example.ecommercebackend.service;
 
+import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.example.ecommercebackend.dto.CartDTO;
 import com.example.ecommercebackend.exceptions.ResourceNotFoundException;
 import com.example.ecommercebackend.models.Cart;
 import com.example.ecommercebackend.models.Products;
 import com.example.ecommercebackend.models.Users;
 import com.example.ecommercebackend.repositories.CartRepository;
+import com.example.ecommercebackend.security.JWTUtility;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
@@ -18,11 +20,14 @@ public class CartService {
 
     private final ProductsService productsService;
 
+    private final JWTUtility jwtUtility;
+
     private final UsersService usersService;
 
-    public CartService(CartRepository cartRepo, ProductsService productsService, UsersService usersService) {
+    public CartService(CartRepository cartRepo, ProductsService productsService, JWTUtility jwtUtility, UsersService usersService) {
         this.cartRepo = cartRepo;
         this.productsService = productsService;
+        this.jwtUtility = jwtUtility;
         this.usersService = usersService;
     }
 
@@ -30,36 +35,52 @@ public class CartService {
         return cartRepo.findAll(Sort.by(Sort.Direction.ASC, "id"));
     }
 
-    public Cart addToCart(CartDTO cartDTO) {
-        Cart cart = new Cart();
+    public List<Cart> getCartItemsByUser(String token) {
+        String tokenUsername = jwtUtility.validateToken(token);
+        Users user = (Users) usersService.loadUserByUsername(tokenUsername);
+        List <Cart> userCartItems = cartRepo.findAllByUsersOrderByDateUpdatedDesc(user);
+        return userCartItems;
+    }
+
+    public Cart addToCart(String token, CartDTO cartDTO) {
         Products foundProducts = productsService.getProductsById(cartDTO.getProducts_id());
-        if (cartDTO.getUsers_id() != 0){
-            Users foundUsers = usersService.getUsersById(cartDTO.getUsers_id());
-            cart.setUsers_id(foundUsers);
-        }
+
+        String tokenUsername = jwtUtility.validateToken(token);
+        Users user = (Users) usersService.loadUserByUsername(tokenUsername);
+
+        Cart cart = new Cart();
+        cart.setUsers(user);
         cart.setProduct_id(foundProducts);
         cart.setQuantity(cartDTO.getQuantity());
-        cart.setDate_added(cartDTO.getDate_added());
-        cart.setDate_updated(cartDTO.getDate_updated());
+        cart.setDateAdded(cartDTO.getDateAdded());
+        cart.setDateUpdated(cartDTO.getDateUpdated());
 
         return cartRepo.save(cart);
     }
 
-    public Cart updateCart(int id, CartDTO cartDTO) {
+    public Cart updateCart(String token, int id, CartDTO cartDTO) {
         Cart foundCart = cartRepo.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Cart not found."));
+        String tokenUsername = jwtUtility.validateToken(token);
+        Users user = (Users) usersService.loadUserByUsername(tokenUsername);
+        if(!user.equals(foundCart.getUsers())){
+            throw new ResourceNotFoundException("Token user does not match cart user");
+        }
         foundCart.setQuantity(cartDTO.getQuantity());
-        foundCart.setDate_updated(cartDTO.getDate_updated());
+        foundCart.setDateUpdated(cartDTO.getDateUpdated());
         Cart updateCart = cartRepo.save(foundCart);
         return updateCart;
     }
 
-    public Cart deleteFromCart(int id) {
+    public Cart deleteFromCart(String token, int id) {
         Cart cart = cartRepo.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Item not found"));
+        String tokenUsername = jwtUtility.validateToken(token);
+        String user = cart.getUsers().getUsername();
+        if (!tokenUsername.equals(user)) {
+            throw new ResourceNotFoundException("Invalid token");
+        }
         cartRepo.delete(cart);
         return cart;
     }
-
-
 }
